Move ControlSwitchButton to a better home
